roll back changes to ScenarioSwitcher
